backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
cryptography
python-multipart
starlette
itsdangerous
